# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Beast {
  id: number;
  name: string;
  /** 1-9 plus 10=Legendary, 11=Dark Legendary */
  tier: number;
  /** Vietnamese name of the tier */
  tierName: string;
  speciesId: number;
  speciesName: string;
  /** Epic, Legend, Mythology, Divine, Folklore, Fantasy */
  raceCategory: string;
  /** Power level of the beast */
  power: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  parentAId?: number | null;
  /** @nullable */
  parentBId?: number | null;
  /**
     * JSON array of abilities
     * @nullable
     */
  abilities?: string | null;
  /** Number of beasts under management */
  subordinateCount?: number;
  isLegendary?: boolean;
  isDarkLegendary?: boolean;
  /** @nullable */
  mapZone?: string | null;
  createdAt: string;
}

export interface BeastListResponse {
  beasts: Beast[];
  total: number;
  page: number;
  limit: number;
}

export interface BeastInput {
  name: string;
  tier: number;
  speciesId: number;
  raceCategory: string;
  power: number;
  description?: string;
  imageUrl?: string;
  abilities?: string;
  mapZone?: string;
}

export interface BeastUpdate {
  name?: string;
  tier?: number;
  power?: number;
  description?: string;
  imageUrl?: string;
  abilities?: string;
  mapZone?: string;
}

export interface BreedingRecord {
  id: number;
  parentAId: number;
  /** @nullable */
  parentAName?: string | null;
  parentBId: number;
  /** @nullable */
  parentBName?: string | null;
  /** @nullable */
  offspringId: number | null;
  /** @nullable */
  offspringName?: string | null;
  success: boolean;
  /** @nullable */
  resultTier?: number | null;
  createdAt: string;
}

export interface BreedingInput {
  parentAId: number;
  parentBId: number;
  offspringName: string;
}

export interface BreedingResult {
  record: BreedingRecord;
  offspring: Beast;
  success: boolean;
  message?: string;
}

export interface Species {
  id: number;
  name: string;
  /** Epic, Legend, Mythology, Divine, Folklore, Fantasy */
  category: string;
  /** Cultural or mythological origin */
  origin: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  beastCount?: number;
}

export interface SpeciesInput {
  name: string;
  category: string;
  origin: string;
  description: string;
  imageUrl?: string;
}

export interface SpeciesUpdate {
  name?: string;
  category?: string;
  origin?: string;
  description?: string;
  imageUrl?: string;
}

export interface UniverseMap {
  id: number;
  tier: string;
  tierName: string;
  zoneName: string;
  description: string;
  climate: string;
  /** 1-10 danger rating */
  dangerLevel: number;
  beastCount?: number;
  /**
     * JSON array of landmark names
     * @nullable
     */
  landmarks?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface StatsOverview {
  totalBeasts: number;
  totalSpecies: number;
  totalBreedings: number;
  legendaryCount: number;
  darkLegendaryCount: number;
  averagePower: number;
  highestTierBeast?: Beast;
}

export interface TierStat {
  tier: number;
  tierName: string;
  count: number;
  averagePower: number;
}

export interface SpeciesStat {
  category: string;
  count: number;
}

export type ListBeastsParams = {
/**
 * @nullable
 */
tier?: number | null;
/**
 * @nullable
 */
species?: string | null;
/**
 * @nullable
 */
raceCategory?: string | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListRecentBeastsParams = {
/**
 * @nullable
 */
limit?: number | null;
};

export type ListBreedingRecordsParams = {
/**
 * @nullable
 */
page?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListSpeciesParams = {
/**
 * @nullable
 */
category?: string | null;
};