# lib/api-zod/src/generated/types/beast.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */

export interface Beast {
  id: number;
  name: string;
  /** 1-9 plus 10=Legendary, 11=Dark Legendary */
  tier: number;
  /** Vietnamese name of the tier */
  tierName: string;
  speciesId: number;
  speciesName: string;
  /** Epic, Legend, Mythology, Divine, Folklore, Fantasy */
  raceCategory: string;
  /** Power level of the beast */
  power: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  parentAId?: number | null;
  /** @nullable */
  parentBId?: number | null;
  /**
     * JSON array of abilities
     * @nullable
     */
  abilities?: string | null;
  /** Number of beasts under management */
  subordinateCount?: number;
  isLegendary?: boolean;
  isDarkLegendary?: boolean;
  /** @nullable */
  mapZone?: string | null;
  createdAt: string;
}