# lib/api-zod/src/generated/types/beastUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */

export interface BeastUpdate {
  name?: string;
  tier?: number;
  power?: number;
  description?: string;
  imageUrl?: string;
  abilities?: string;
  mapZone?: string;
}