# lib/api-zod/src/generated/types/breedingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */

export interface BreedingInput {
  parentAId: number;
  parentBId: number;
  offspringName: string;
}