# lib/api-zod/src/generated/types/breedingRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */

export interface BreedingRecord {
  id: number;
  parentAId: number;
  /** @nullable */
  parentAName?: string | null;
  parentBId: number;
  /** @nullable */
  parentBName?: string | null;
  /** @nullable */
  offspringId: number | null;
  /** @nullable */
  offspringName?: string | null;
  success: boolean;
  /** @nullable */
  resultTier?: number | null;
  createdAt: string;
}