# lib/api-zod/src/generated/types/breedingResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */
import type { Beast } from './beast';
import type { BreedingRecord } from './breedingRecord';

export interface BreedingResult {
  record: BreedingRecord;
  offspring: Beast;
  success: boolean;
  message?: string;
}