# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */
import type { Beast } from './beast';

export interface StatsOverview {
  totalBeasts: number;
  totalSpecies: number;
  totalBreedings: number;
  legendaryCount: number;
  darkLegendaryCount: number;
  averagePower: number;
  highestTierBeast?: Beast;
}