# lib/api-zod/src/generated/types/tierStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */

export interface TierStat {
  tier: number;
  tierName: string;
  count: number;
  averagePower: number;
}