# lib/api-zod/src/generated/types/universeMap.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thú Giới API - Animal Universe
 * OpenAPI spec version: 0.1.0
 */

export interface UniverseMap {
  id: number;
  tier: string;
  tierName: string;
  zoneName: string;
  description: string;
  climate: string;
  /** 1-10 danger rating */
  dangerLevel: number;
  beastCount?: number;
  /**
     * JSON array of landmark names
     * @nullable
     */
  landmarks?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}